backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Esteban Glochon",email = "esteban.glochon@usach.cl"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
